Merged PR 152: fix numerical head in distributed training
Fix numerical head in distributed training. Previously, the `num_head` is used in forward pass only when there's numerical labels in the batch. This can cause problems when distributed training (not very sure which mode causes the problem, ddp, zero or pp), because some ranks may have gradients of the parameters of `num_head`, while others may not. The communication between ranks may get stuck due to the difference in gradient shape.