Merged PR 254: chore(SFM): build sdist package to prepare for release
* select files which need distributing via MANIFEST.in, add check for incomplete manifest
* determine package version based on most recent git tag (c.f. git-describe)
* name builds based on branch or tag names
* build wheel + sdist file, and upload them as artifacts if the pipeline is a scheduled/tag build

Once this is merged, we can create a release pipeline that takes the artifact and uploads it e.g. to the artifacts feed.

Note that we could do something similar for docs, ie, build them and upload them as artifacts. This could be done in the same or a separate pipeline.