Merged PR 292: Feat(Sample): Implement sample for angle
# What's new
+ Add the script `amlt/tox/run_local_sample.sh` to sample locally
+ Delete `scripts/tox/sample_tox_angle.sh` and `tasks/tox/sample_tox_angle.py` but migrate to `scripts/tox/sample_tox.sh` and `tasks/tox/sample_tox.py`
+ Modify `tasks/tox/sample_tox.py` and `Tox.sample` function in `sfm/models/toxmodel.py` to implement sampling correctly.
+ Add one argument `t_total_steps` in `scripts/tox/sample_tox.py` that denotes the steps for solving the sde integral.
+ Multiple GPUs can be used for sampling

# Suggestion
If the argument is bool and you want it to be loaded from an environment variable and then passed into the python startup command, use the following statement.
```shell
$("$add_3d" == "true" && echo "--add_3d")
```
I only use it in `scripts/tox/sample_tox.sh`, if you need to use it in all scripts, you need to raise another pull request to do this.