Merged PR 567: add dpm-edm solver for fast sampling, 5-10 steps are all you need
**Usage**: set the following variable to dpm_edm and edm

`diffusion_sampling=dpm_edm`
`diffusion_mode=edm`

Must be used with edm model. 5 steps are enough depicted by experiments with the setting
`num_timesteps_stepsize=-1000`
when using
`num_timesteps=5000`, i.e., 5 = 5000/1000.

To be cautious, 10 steps are encouraged.

Ps: this pr is for sampling of edm model, most of the training scripts are not updated except for some codes used in sampling.